+++ src/main/java/com/sun/caresyncsystem/utils/api/BookingApiPaths.java
@@ -2,8 +2,4 @@ package com.sun.caresyncsystem.utils.api;
 
 public class BookingApiPaths {
     public static final String BASE = "/api/v1/booking";
-
-    public static class Endpoint {
-        public static final String BY_ID = "/{id}";
-    }
 }
