+++ src/main/java/com/sun/caresyncsystem/utils/api/DoctorApiPaths.java
@@ -6,5 +6,6 @@ public class DoctorApiPaths {
     public static class Endpoint {
         public static final String SEARCH = "/search";
         public static final String FULL_SEARCH = BASE + SEARCH;
+        public static final String BOOKING_BY_ID = "/bookings/{id}";
     }
 }
